Hafi - update thread bagia user

+++ controllers/main/c_thread.js
@@ -1,6 +1,8 @@
 const Thread = require('../../models/m_thread');
 const Comment = require('./../../models/m_comment');
+const env = require('dotenv');
 const {res_error, res_success} = require('../../response')
+env.config();
 
 const getAllThreads = async (req, res) => {
     try {
@@ -48,7 +50,7 @@ const updateThreadById = async (req, res) => {
         let id_thread = req.params.id;
         let data = req.body;
         
-        if(req.user.user._id != data.user) return res_error(res, 403, "403 Forbidden", "Unauthenticated error and incorrect address so can't delete thread by id");
+        if(req.user.user._id != data.user || req.user.user.role != process.env.ADMIN) return res_error(res, 403, "403 Forbidden", "Unauthenticated error and incorrect address so can't delete thread by id");
 
         await Thread.updateOne({"_id":id_thread}, {$set:data}, (err, result) => {
             if(err) return res_error(res, 400, "400 Bad Request", "Request error by client so that it cannot change thread by ID")
@@ -65,7 +67,7 @@ const deleteThread = async (req, res) => {
         let id_thread = req.params.id;
         let {user} = req.body;
 
-        if(req.user.user._id != user) return res_error(res, 403, "403 Forbidden", "Unauthenticated error and incorrect address so can't delete thread by id");
+        if(req.user.user._id != user || req.user.user.role != process.env.ADMIN) return res_error(res, 403, "403 Forbidden", "Unauthenticated error and incorrect address so can't delete thread by id");
 
         await Comment.deleteMany({"thread":id_thread});
